Keysize algorithm fix and refactoring
Instead of strictly picking the "best" normalized hamming distance
(i.e. a decimal - 2.345124), I decided to cast the normalized value to
an int and group the keylengths. The reason I did this was because
if a hamming distance defines the number of bits that have to change
between two strings, there's no such thing as "fractional" bits - therefore
it seems reasonable to use the integer portion of the keysizes. An argument
could be made for using rounding instead of integer casting, but as it stands,
this approach may not be perfect, but satisfies being "good enough."